converted Notify to lazy component
