+++ client/src/pages/Notify/index.tsx
@@ -1,18 +1,3 @@
-import { useNotify } from 'hooks/useNotify';
-import { Button, Container } from 'react-bootstrap';
+import { lazy } from 'react';
 
-export const Notify = () => {
-  const { permission, requestPermission, createNotification } = useNotify();
-  const testNotification = () => {
-    if (permission !== 'granted') requestPermission();
-
-    createNotification('time', 'name');
-  };
-
-  return (
-    <Container>
-      <h2>Test if notifications work here</h2>
-      <Button onClick={testNotification}>Click me</Button>
-    </Container>
-  );
-};
+export const NotifyLazy = lazy(() => import('./Notify'));
