+++ client/src/pages/index.ts
@@ -4,4 +4,3 @@ export * from './Home';
 export * from './Medicine';
 export * from './Medicines';
 export * from './NotFound';
-export * from './Notify';
